made reader into it's own thing
